requirements.txt
python-dotenv
requests
schedule
coinbase-advanced-py
